requirements.txt
pyTelegramBotAPI
fs-pochta-api